Refactor excel/conversion functions to excelUtils
